Join metadata and Nextclade on explicit columns
Replaces a Snakemake `run` block based on pandas and an assumption that
the metadata and Nextclade tables contain the same strains in the same
order with a `shell` block based on tsv-utils and csvtk commands that
select a user-defined list of columns from the Nextclade input and join
the inputs on the specific id columns.

Potentially controversial is the loss of lineage-specific clade column
originally provided by this rule (e.g., supporting "short_clade" only
for Vic). This functionality could be added back downstream from this
join by allowing a lineage-specific "frequency category" for the
frequency estimation rules.