Parameterize clades to plot, Auspice config, and coloring field
Adds optional parameters to the multi-region plot rule allowing users to
define clades to plot, an Auspice config JSON with a color scale to use,
and the name of the config JSON's coloring field whose scale should be
used. Updates the flu config to define the specific subclades to plot
per lineage and the Auspice config JSONs to use. Colors the frequencies
with the "subclade" color scale in the config JSONs.